test: consolidate earliest_available tests into one file
Move the resharding and state-sync earliest_available regression tests into `tests/earliest_available.rs`; expose `sync::util` as `pub(crate)` so the state-sync test can reuse the far-horizon helpers.
