+++ test-loop-tests/src/tests/sync/mod.rs
@@ -1,5 +1,4 @@
 mod continuous_epoch_sync;
-mod earliest_available;
 mod epoch_sync;
 mod far_horizon;
 mod gc;
@@ -8,5 +7,5 @@ mod near_horizon;
 mod state_sync;
 mod sync_then_catchup;
 mod syncing;
-mod util;
+pub(crate) mod util;
 mod validator_kickout;
